// cgoml-ui/src/app/features/models/messaging/filters/InboundMessageFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class InboundMessageFilter {

uuid: string;
receptionDateMinValue: Date;
receptionDateMaxValue: Date;
messageType: string;
status: string;

/* Specific Code Start */
/* Specific Code End */
}