// cgoml-ui/src/app/features/models/messaging/sortings/InboundMessageSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class InboundMessageSorting {

uuidOrderType: string;
receptionDateOrderType: string;
messageTypeOrderType: string;
statusOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}